Moved classes/interfaces into appropriate packages.
* The `Game` package contains general game classes. It includes
  the game scenes drawn by the GameSceneManager as well as the
  GameView which exposes a canvas that can be drawn on.
* The `Gameplay` package contains classes used for the core
  game play: playing with the motorcycle.
* The `Math` package contains classes used for Physics and
  collision detection.
