+++ BrokenBonez/app/src/main/java/com/dragonfruitstudios/brokenbonez/GameActivity.java
@@ -3,13 +3,12 @@ package com.dragonfruitstudios.brokenbonez;
 import android.app.Activity;
 import android.os.Bundle;
 import android.os.PowerManager;
-import android.util.Log;
 import android.view.KeyEvent;
 import android.view.MotionEvent;
 import android.view.WindowManager;
 
 import com.dragonfruitstudios.brokenbonez.AssetLoading.AssetLoader;
-import com.dragonfruitstudios.brokenbonez.Input.TouchHandler;
+import com.dragonfruitstudios.brokenbonez.Game.GameView;
 
 /**
  * Game Activity class used for creating a new game view and game loop instance. Also defines some
