+++ BrokenBonez/app/src/main/java/com/dragonfruitstudios/brokenbonez/GameSceneManager.java
@@ -2,7 +2,8 @@ package com.dragonfruitstudios.brokenbonez;
 
 import android.util.Log;
 
-import com.dragonfruitstudios.brokenbonez.Scenes.Scene;
+import com.dragonfruitstudios.brokenbonez.Game.GameView;
+import com.dragonfruitstudios.brokenbonez.Game.Scenes.Scene;
 
 import java.util.HashMap;
